school-ai-teacher/backend/requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
PyPDF2
python-docx
beautifulsoup4
requests
